crc-demo-main/requirements.txt
streamlit>=1.28.0
torch>=2.5.0
torchvision>=0.20.0
pillow>=10.2.0
numpy>=1.24.0,<2.0.0
plotly>=5.10.0
pandas>=2.0.0
matplotlib>=3.8.0
opencv-python-headless>=4.5.0
reportlab>=4.0.0
seaborn>=0.13.0
flask>=2.0.0
lifelines>=0.27.0
scikit-learn>=1.3.0
tqdm>=4.65.0
scipy>=1.9.0
networkx>=2.8.0
scikit-image>=0.19.0
protobuf>=3.20,<5
psutil>=5.9.0
# PyRadiomics and hybrid classifier dependencies
pyradiomics>=3.0.1
SimpleITK>=2.1.0
xgboost>=1.7.0
shap>=0.41.0
boruta>=0.3
feature-engine>=1.6.0 